+++ src/pages/MakeSurbeySiwperPage.js
@@ -1,47 +1,9 @@
-import React, { useState } from "react";
-import { Swiper, SwiperSlide } from "swiper/react";
-import SwiperCore, { Navigation, Pagination } from "swiper";
-import "swiper/css"; //basic
-import "swiper/css/navigation";
-import "swiper/css/pagination";
-import { useDispatch, useSelector } from "react-redux";
-import EditQuestionDetail from "../components/editQuestionDetail";
-import EditAnswerDetail from "../components/editAnswerDetail";
-import { editAnswer, editQuestion, editTime } from "../modules/editQuestion";
-import { useLocation } from "react-router-dom";
+import React from "react";
+import { Link, useLocation } from "react-router-dom";
+import MakeSurbeySwiperContainer from "../containers/MakeSurbeySwiperContainer";
 
-SwiperCore.use([Navigation, Pagination]);
-
-const outerStyle = {
-  height: "15em",
-  fontSize: "1rem",
-  lineHeight: 1.5,
-  display: "flex",
-};
-
-const innerStyle = {
-  width: "10em",
-  border: "1px solid teal",
-  borderRadius: ".25rem",
-  margin: "0 auto",
-};
-
-function MakeSurbeySwiperPage({ index }) {
-  const dispatch = useDispatch();
+function MakeSurbeySwiperPage() {
   const location = useLocation();
-
-  const [timeCount, setTimeCount] = useState(0);
-
-  const editQuestions = useSelector((state) => state.editReducer);
-  const onEdit = (id, text) => dispatch(editQuestion(id, text)); //질문 수정
-  const onEditTime = (id, time) => dispatch(editTime(id, time)); //질문 수정
-  const onEditAnswer = (questionId, answerId, text) =>
-    dispatch(editAnswer(questionId, answerId, text)); //답변 수정
-
-  const onChange = (e) => {
-    setTimeCount(1 - timeCount);
-    onEditTime(parseInt(e.target.name), timeCount);
-  };
 
   const focusIndex =
     location.state.focusIndex == null ? 1 : location.state.focusIndex;
@@ -50,49 +12,10 @@ function MakeSurbeySwiperPage({ index }) {
 
   return (
     <div>
-      <Swiper
-        className="banner"
-        spaceBetween={50}
-        slidesPerView={1}
-        initialSlide={focusIndex}
-        navigation
-        allowTouchMove={false}
-        pagination={{ clickable: true }}
-      >
-        {editQuestions &&
-          editQuestions.map((question, index) => (
-            <SwiperSlide>
-              <div key={question.id} style={outerStyle}>
-                <div style={innerStyle}>
-                  <div> Q{index} </div>
-                  <EditQuestionDetail
-                    question={question}
-                    onEdit={onEdit}
-                    id={question.id}
-                  />
-
-                  <button name={question.id} onClick={onChange}>
-                    {timeCount ? "타이머 끄기" : "타이머 켜기"}
-                  </button>
-
-                  <EditAnswerDetail
-                    answer={question.answer[0].text}
-                    onEdit={onEditAnswer}
-                    questionId={question.id}
-                    answerId={0}
-                  />
-
-                  <EditAnswerDetail
-                    answer={question.answer[1].text}
-                    onEdit={onEditAnswer}
-                    questionId={question.id}
-                    answerId={1}
-                  />
-                </div>
-              </div>
-            </SwiperSlide>
-          ))}
-      </Swiper>
+      <div>
+        <Link to="/edit"> 뒤로가기 </Link>
+      </div>
+      <MakeSurbeySwiperContainer focusIndex={focusIndex} />
     </div>
   );
 }
